@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.18)
 
 include(GoogleTest)
 aux_source_directory(../src  DIR_SRCS)
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 17)
 
 file(GLOB_RECURSE NET_TEST_SOURCE "${PROJECT_SOURCE_DIR}/test/*.cc")
 
