@@ -12,7 +12,7 @@
 
 class PikaAuxiliaryThread : public net::Thread {
  public:
-  PikaAuxiliaryThread() : mu_(), cv_(&mu_) { set_thread_name("AuxiliaryThread"); }
+  PikaAuxiliaryThread() { set_thread_name("AuxiliaryThread"); }
   virtual ~PikaAuxiliaryThread();
   pstd::Mutex mu_;
   pstd::CondVar cv_;
