@@ -5,6 +5,8 @@
 
 #include "net/include/bg_thread.h"
 #include <sys/time.h>
+#include <cstdlib>
+#include <mutex>
 
 #include "pstd/include/pstd_mutex.h"
 #include "pstd/include/xdebug.h"
@@ -12,99 +14,92 @@
 namespace net {
 
 void BGThread::Schedule(void (*function)(void*), void* arg) {
-  mu_.Lock();
-  while (queue_.size() >= full_ && !should_stop()) {
-    wsignal_.Wait();
-  }
+  std::unique_lock lock(mu_);
+
+  wsignal_.wait(lock, [this]() { return queue_.size() < full_ || should_stop(); });
+
   if (!should_stop()) {
-    queue_.push(BGItem(function, arg));
-    rsignal_.Signal();
+    queue_.emplace(function, arg);
+    rsignal_.notify_one();
   }
-  mu_.Unlock();
 }
 
 void BGThread::QueueSize(int* pri_size, int* qu_size) {
-  pstd::MutexLock l(&mu_);
+  std::lock_guard lock(mu_);
   *pri_size = timer_queue_.size();
   *qu_size = queue_.size();
 }
 
 void BGThread::QueueClear() {
-  pstd::MutexLock l(&mu_);
+  std::lock_guard lock(mu_);
   std::queue<BGItem>().swap(queue_);
   std::priority_queue<TimerItem>().swap(timer_queue_);
-  wsignal_.Signal();
+  wsignal_.notify_one();
 }
 
 void BGThread::SwallowReadyTasks() {
   // it's safe to swallow all the remain tasks in ready and timer queue,
   // while the schedule function would stop to add any tasks.
-  mu_.Lock();
+  mu_.lock();
   while (!queue_.empty()) {
-    void (*function)(void*) = queue_.front().function;
-    void* arg = queue_.front().arg;
+    auto [function, arg] = queue_.front();
     queue_.pop();
-    mu_.Unlock();
+    mu_.unlock();
     (*function)(arg);
-    mu_.Lock();
+    mu_.lock();
   }
-  mu_.Unlock();
+  mu_.unlock();
 
-  struct timeval now;
-  gettimeofday(&now, nullptr);
-  uint64_t unow = now.tv_sec * 1000000 + now.tv_usec;
-  mu_.Lock();
+  auto now = std::chrono::system_clock::now();
+  uint64_t unow = std::chrono::duration_cast<std::chrono::microseconds>(now.time_since_epoch()).count();
+  mu_.lock();
+
   while (!timer_queue_.empty()) {
-    TimerItem top_item = timer_queue_.top();
-    if (unow / 1000 < top_item.exec_time / 1000) {
+    auto [exec_time, function, arg] = timer_queue_.top();
+    if (unow < exec_time) {
       break;
     }
-    void (*function)(void*) = top_item.function;
-    void* arg = top_item.arg;
     timer_queue_.pop();
     // Don't lock while doing task
-    mu_.Unlock();
+    mu_.unlock();
     (*function)(arg);
-    mu_.Lock();
+    mu_.lock();
   }
-  mu_.Unlock();
+  mu_.unlock();
 }
 
 void* BGThread::ThreadMain() {
   while (!should_stop()) {
-    mu_.Lock();
-    while (queue_.empty() && timer_queue_.empty() && !should_stop()) {
-      rsignal_.Wait();
-    }
+    std::unique_lock lock(mu_);
+
+    rsignal_.wait(lock, [this]() { return !queue_.empty() || !timer_queue_.empty() || should_stop(); });
+
     if (should_stop()) {
-      mu_.Unlock();
       break;
     }
-    if (!timer_queue_.empty()) {
-      struct timeval now;
-      gettimeofday(&now, nullptr);
 
-      TimerItem timer_item = timer_queue_.top();
-      uint64_t unow = now.tv_sec * 1000000 + now.tv_usec;
-      if (unow / 1000 >= timer_item.exec_time / 1000) {
-        void (*function)(void*) = timer_item.function;
-        void* arg = timer_item.arg;
+    if (!timer_queue_.empty()) {
+      auto now = std::chrono::system_clock::now();
+      uint64_t unow = std::chrono::duration_cast<std::chrono::microseconds>(now.time_since_epoch()).count();
+      auto [exec_time, function, arg] = timer_queue_.top();
+      if (unow >= exec_time) {
         timer_queue_.pop();
-        mu_.Unlock();
+        lock.unlock();
         (*function)(arg);
         continue;
       } else if (queue_.empty() && !should_stop()) {
-        rsignal_.TimedWait(static_cast<uint32_t>((timer_item.exec_time - unow) / 1000));
-        mu_.Unlock();
+        rsignal_.wait_for(lock, std::chrono::microseconds(exec_time - unow));
+
+        lock.unlock();
         continue;
       }
     }
+
     if (!queue_.empty()) {
-      void (*function)(void*) = queue_.front().function;
-      void* arg = queue_.front().arg;
+      auto [function, arg] = queue_.front();
       queue_.pop();
-      wsignal_.Signal();
-      mu_.Unlock();
+      wsignal_.notify_one();
+      lock.unlock();
       (*function)(arg);
     }
   }
@@ -117,21 +112,15 @@ void* BGThread::ThreadMain() {
  * timeout is in millisecond
  */
 void BGThread::DelaySchedule(uint64_t timeout, void (*function)(void*), void* arg) {
-  /*
-   * pthread_cond_timedwait api use absolute API
-   * so we need gettimeofday + timeout
-   */
-  struct timeval now;
-  gettimeofday(&now, nullptr);
-  uint64_t exec_time;
-  exec_time = now.tv_sec * 1000000 + timeout * 1000 + now.tv_usec;
+  auto now = std::chrono::system_clock::now();
+  uint64_t unow = std::chrono::duration_cast<std::chrono::microseconds>(now.time_since_epoch()).count();
+  uint64_t exec_time = unow + timeout * 1000;
 
-  mu_.Lock();
+  std::lock_guard lock(mu_);
   if (!should_stop()) {
-    timer_queue_.push(TimerItem(exec_time, function, arg));
-    rsignal_.Signal();
+    timer_queue_.emplace(exec_time, function, arg);
+    rsignal_.notify_one();
   }
-  mu_.Unlock();
 }
 
 }  // namespace net
