@@ -45,10 +45,10 @@ NetItem NetMultiplexer::NotifyQueuePop() {
   }
 
   NetItem it;
-  notify_queue_protector_.Lock();
+  notify_queue_protector_.lock();
   it = notify_queue_.front();
   notify_queue_.pop();
-  notify_queue_protector_.Unlock();
+  notify_queue_protector_.unlock();
   return it;
 }
 
@@ -59,12 +59,12 @@ bool NetMultiplexer::Register(const NetItem& it, bool force) {
   }
 
   bool success = false;
-  notify_queue_protector_.Lock();
+  notify_queue_protector_.lock();
   if (force || queue_limit_ == kUnlimitedQueue || notify_queue_.size() < static_cast<size_t>(queue_limit_)) {
     notify_queue_.push(it);
     success = true;
   }
-  notify_queue_protector_.Unlock();
+  notify_queue_protector_.unlock();
   if (success) {
     write(notify_send_fd_, "", 1);
   }
