@@ -115,7 +115,7 @@ ReadStatus PbConn::GetRequest() {
 WriteStatus PbConn::SendReply() {
   ssize_t nwritten = 0;
   size_t item_len;
-  pstd::MutexLock l(&resp_mu_);
+  std::lock_guard l(resp_mu_);
   while (!write_buf_.queue_.empty()) {
     std::string item = write_buf_.queue_.front();
     item_len = item.size();
@@ -147,7 +147,7 @@ WriteStatus PbConn::SendReply() {
 }
 
 void PbConn::set_is_reply(const bool is_reply) {
-  pstd::MutexLock l(&is_reply_mu_);
+  std::lock_guard l(is_reply_mu_);
   if (is_reply) {
     is_reply_++;
   } else {
@@ -159,14 +159,14 @@ void PbConn::set_is_reply(const bool is_reply) {
 }
 
 bool PbConn::is_reply() {
-  pstd::MutexLock l(&is_reply_mu_);
+  std::lock_guard l(is_reply_mu_);
   return is_reply_ > 0;
 }
 
 int PbConn::WriteResp(const std::string& resp) {
   std::string tag;
   BuildInternalTag(resp, &tag);
-  pstd::MutexLock l(&resp_mu_);
+  std::lock_guard l(resp_mu_);
   write_buf_.queue_.push(tag);
   write_buf_.queue_.push(resp);
   set_is_reply(true);
